Fix a bug in SimpleWorkloadDriver related with json parsing
This bug led to false "Finished diagnosis successfully" messages in the
UI if a ClassNotFoundException had been thrown (e.g. user configured
non-existing VU script class) during a failed diagnosis because this
class cannot be serialized and transmitted correctly using the rest
service.

The reason behind this is that the class ClassNotFoundException uses its
own field "ex" instead of "cause" of the super class Throwable. It has
for legacy reasons two methods to access this field, the old one
"getException()" and the new one "getCause()". Now when serializing an
object of this class, the jackson json processor struggles with the
field "ex" and throws an UnrecognizedPropertyException. Thus the
exception cannot be included in the tree and had to be replaced by an
according message instead.

Change-Id: I59057ae61296a95ddbc437b4b7aa61174c0ee70c
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>